Font related refactoring
Mainly refactoring, not real and wow changes

+++ src/com/sun/pdfview/font/ttf/HeadTable.java
@@ -97,9 +97,9 @@ public class HeadTable extends TrueTypeTable {
         setCreated(System.currentTimeMillis());
         setModified(System.currentTimeMillis());
         setXMin((short) 0);
-        setXMax(Short.MAX_VALUE);
+        setXMax((short) Short.MAX_VALUE);
         setYMin((short) 0);
-        setYMax(Short.MAX_VALUE);
+        setYMax((short) Short.MAX_VALUE);
         setMacStyle((short) 0x0);
         setLowestRecPPem((short) 0);
         setFontDirectionHint((short) 0);
